+++ src/grammar/expression_parser.rs
@@ -13,6 +13,11 @@ use crate::types::{
 };
 use std::cmp::Ordering;
 
+// Local-timezone handling for `now` and bare times lives in a child module so it
+// can access `ExpressionParser`'s private fields while keeping this file small.
+#[path = "expression_parser_timezone.rs"]
+mod timezone;
+
 /// Evaluates a power expression, using exact rational arithmetic when possible.
 ///
 /// When both base and exponent are rational and the exponent is an integer
@@ -90,28 +95,6 @@ impl ExpressionParser {
             currency_db: CurrencyDatabase::new(),
             current_date_context: None,
             local_offset_seconds: None,
-        }
-    }
-
-    /// Sets the user's local timezone offset in seconds east of UTC.
-    ///
-    /// Pass `None` to fall back to the default UTC interpretation.
-    pub fn set_local_offset_seconds(&mut self, offset_seconds: Option<i32>) {
-        self.local_offset_seconds = offset_seconds;
-    }
-
-    /// Returns the configured local timezone offset in seconds east of UTC, if any.
-    #[must_use]
-    pub fn local_offset_seconds(&self) -> Option<i32> {
-        self.local_offset_seconds
-    }
-
-    /// Returns a `DateTime` representing the current instant, honoring the
-    /// configured local timezone offset when one is set.
-    fn current_now(&self) -> DateTime {
-        match self.local_offset_seconds {
-            Some(offset) => DateTime::now_local(offset),
-            None => DateTime::now_with_label("current UTC time", Some(0), Some("UTC".to_string())),
         }
     }
